+++ tests/testthat/test-bcbioSingleCell.R
@@ -1,24 +1,36 @@
 context("bcbioSingleCell")
 
-test_that("bcbioSingleCell", {
-    uploadDir <- system.file("extdata/indrops", package = "bcbioSingleCell")
+uploadDir <- system.file("extdata/indrops", package = "bcbioSingleCell")
 
-    ## Minimal mode, with no metadata or annotations.
-    ## This is fast but doesn't slot a lot of useful info.
+## Minimal mode, with no metadata or annotations.
+## This is fast but doesn't slot a lot of useful info.
+test_that("Minimal mode", {
     x <- bcbioSingleCell(uploadDir = uploadDir)
     expect_s4_class(x, "bcbioSingleCell")
+})
 
-    ## User-defined metadata.
+test_that("User-defined metadata", {
     x <- bcbioSingleCell(
         uploadDir = uploadDir,
         sampleMetadataFile <- file.path(uploadDir, "metadata.csv")
     )
     expect_s4_class(x, "bcbioSingleCell")
+})
 
-    ## Automatic organism annotations from AnnotationHub.
+## Automatic organism annotations from AnnotationHub.
+test_that("AnnotationHub", {
     x <- bcbioSingleCell(
         uploadDir = uploadDir,
         organism = "Homo sapiens"
     )
     expect_s4_class(x, "bcbioSingleCell")
+})
+
+## Setting to serial by default, for cross-platform compatibility.
+test_that("BPPARAM", {
+    skip_on_appveyor()
+     x <- bcbioSingleCell(
+         uploadDir = uploadDir,
+         BPPARAM = BiocParallel::MulticoreParam()
+     )
 })