+++ tests/testthat/test-plotCellCounts.R
@@ -1,6 +1,0 @@
-context("plotCellCounts")
-
-test_that("bcbioSingleCell", {
-    x <- plotCellCounts(indrops)
-    expect_s3_class(x, "ggplot")
-})