+++ tests/testthat/test-plotMitoVsCoding.R
@@ -1,6 +1,0 @@
-context("plotMitoVsCoding")
-
-test_that("bcbioSingleCell", {
-    x <- plotMitoVsCoding(indrops)
-    expect_s3_class(x, "ggplot")
-})