+++ tests/testthat/test-updateObject.R
@@ -1,12 +1,13 @@
 context("updateObject")
 
 test_that("bcbioSingleCell", {
-    x <- updateObject(indrops)
+    x <- updateObject(bcb)
     expect_s4_class(x, "bcbioSingleCell")
 })
 
 test_that("v0.1 update", {
     skip_if_not(hasInternet())
+    skip_on_appveyor()
     invalid <- import(
         file = file.path(bcbioSingleCellTestsURL, "bcbioSingleCell_0.1.0.rds")
     )